Organize result types
Major breaking changes including abstractifying GriddlyResult<> and
removal of DapperGriddlyResult
