+++ Griddly.Mvc/GriddlyResult.cs
@@ -39,18 +39,13 @@ namespace Griddly.Mvc
         }
     }
 
-    public class GriddlyResult<T> : GriddlyResult
+    public abstract class GriddlyResult<T> : GriddlyResult
     {
-        IQueryable<T> _result;
-        Func<IQueryable<T>, IQueryable<T>> _massage = null;
-
         public string ViewName { get; set; }
 
-        public GriddlyResult(IQueryable<T> result, string viewName = null, Func<IQueryable<T>, IQueryable<T>> massage = null)
+        public GriddlyResult(string viewName = null)
         {
-            _result = result;
             ViewName = viewName;
-            _massage = massage;
         }
 
         public override void ExecuteResult(ControllerContext context)
@@ -166,157 +161,13 @@ namespace Griddly.Mvc
             }
         }
 
-        public virtual IEnumerable<T> GetAll(SortField[] sortFields)
-        {
-            IQueryable<T> sortedQuery = ApplySortFields(_result, sortFields);
+        public abstract IEnumerable<T> GetAll(SortField[] sortFields);
+        
+        public abstract IList<T> GetPage(int pageNumber, int pageSize, SortField[] sortFields);
+        
+        public abstract void PopulateSummaryValues(GriddlySettings<T> settings);
 
-            if (_massage != null)
-                sortedQuery = _massage(sortedQuery);
-
-            return sortedQuery;
-        }
-
-        public virtual IList<T> GetPage(int pageNumber, int pageSize, SortField[] sortFields)
-        {
-            IQueryable<T> sortedQuery = ApplySortFields(_result, sortFields);
-
-            if (_massage != null)
-                sortedQuery = _massage(sortedQuery);
-
-            return sortedQuery.Skip(pageNumber * pageSize).Take(pageSize).ToList();
-        }
-
-        public virtual void PopulateSummaryValues(GriddlySettings<T> settings)
-        {
-            // Only works for linq to objects
-            //List<GriddlyColumn> summaryColumns = settings.Columns.Where(x => x.SummaryFunction != null).ToList();
-
-            //if (summaryColumns.Any())
-            //{
-            //    StringBuilder aggregateExpression = new StringBuilder();
-
-            //    aggregateExpression.Append("new (");
-
-            //    for (int i = 0; i < summaryColumns.Count; i++)
-            //    {
-            //        if (i > 0)
-            //            aggregateExpression.Append(", ");
-
-            //        GriddlyColumn col = summaryColumns[i];
-
-            //        aggregateExpression.AppendFormat("{0}({1}) AS _a{2}", col.SummaryFunction, col.ExpressionString, i);
-            //    }
-
-            //    aggregateExpression.Append(")");
-
-            //    var query = _result.GroupBy(x => 1).Select(aggregateExpression.ToString());
-            //    var item = query.Cast<object>().Single();
-            //    var type = item.GetType();
-
-            //    for (int i = 0; i < summaryColumns.Count; i++)
-            //        summaryColumns[i].SummaryValue = type.GetProperty("_a" + i).GetValue(item);
-            //}
-
-            // TODO: figure out how to get this in one query
-            foreach (GriddlyColumn c in settings.Columns.Where(x => x.SummaryFunction != null))
-            {
-                switch (c.SummaryFunction.Value)
-                {
-                    case SummaryAggregateFunction.Sum:
-                    case SummaryAggregateFunction.Average:
-                    case SummaryAggregateFunction.Min:
-                    case SummaryAggregateFunction.Max:
-                        c.SummaryValue = _result.Aggregate(c.SummaryFunction.Value.ToString(), c.ExpressionString);
-
-                        break;
-
-                    default:
-                        throw new InvalidOperationException(string.Format("Unknown summary function {0} for column {1}.", c.SummaryFunction, c.ExpressionString));
-                }
-            }
-        }
-
-        public virtual long GetCount()
-        {
-            return _result.Count();
-        }
-
-        protected static IQueryable<T> ApplySortFields(IQueryable<T> source, SortField[] sortFields)
-        {
-            IOrderedQueryable<T> sortedQuery = null;
-
-            if (sortFields != null)
-            {
-                for (int i = 0; i < sortFields.Length; i++)
-                {
-                    SortField sortField = sortFields[i];
-
-                    if (sortField.Direction == SortDirection.Ascending)
-                    {
-                        if (i == 0)
-                            sortedQuery = OrderBy(source, sortField.Field);
-                        else
-                            sortedQuery = ThenBy(sortedQuery, sortField.Field);
-                    }
-                    else
-                    {
-                        if (i == 0)
-                            sortedQuery = OrderByDescending(source, sortField.Field);
-                        else
-                            sortedQuery = ThenByDescending(sortedQuery, sortField.Field);
-                    }
-                }
-            }
-
-            return sortedQuery ?? source;
-        }
-
-        // http://stackoverflow.com/a/233505/8037
-        static IOrderedQueryable<T> OrderBy(IQueryable<T> source, string property)
-        {
-            return ApplyOrder(source, property, "OrderBy");
-        }
-
-        static IOrderedQueryable<T> OrderByDescending(IQueryable<T> source, string property)
-        {
-            return ApplyOrder(source, property, "OrderByDescending");
-        }
-
-        static IOrderedQueryable<T> ThenBy(IOrderedQueryable<T> source, string property)
-        {
-            return ApplyOrder(source, property, "ThenBy");
-        }
-
-        static IOrderedQueryable<T> ThenByDescending(IOrderedQueryable<T> source, string property)
-        {
-            return ApplyOrder(source, property, "ThenByDescending");
-        }
-
-        static IOrderedQueryable<T> ApplyOrder(IQueryable<T> source, string property, string methodName)
-        {
-            string[] props = property.Split('.');
-            Type type = typeof(T);
-            ParameterExpression arg = Expression.Parameter(type, "x");
-            Expression expr = arg;
-            foreach (string prop in props)
-            {
-                // use reflection (not ComponentModel) to mirror LINQ
-                PropertyInfo pi = type.GetProperty(prop);
-                expr = Expression.Property(expr, pi);
-                type = pi.PropertyType;
-            }
-            Type delegateType = typeof(Func<,>).MakeGenericType(typeof(T), type);
-            LambdaExpression lambda = Expression.Lambda(delegateType, expr, arg);
-
-            object result = typeof(Queryable).GetMethods().Single(
-                    method => method.Name == methodName
-                            && method.IsGenericMethodDefinition
-                            && method.GetGenericArguments().Length == 2
-                            && method.GetParameters().Length == 2)
-                    .MakeGenericMethod(typeof(T), type)
-                    .Invoke(null, new object[] { source, lambda });
-            return (IOrderedQueryable<T>)result;
-        }
+        public abstract long GetCount();
     }
 
     public enum GriddlyExportFormat
